addd src

+++ build.py
@@ -71,9 +71,10 @@ def buildSip():
                 per = 100
             print('download sip-%s.tar.gz %.2f%%' % (args.sipver, per))
 
-        urllib.request.urlretrieve(
-            'https://www.riverbankcomputing.com/static/Downloads/sip/{0}/sip-{0}.tar.gz'.format(args.sipver),
-            path, reporthook)
+        if not os.path.isfile(path):
+            urllib.request.urlretrieve(
+                'https://www.riverbankcomputing.com/static/Downloads/sip/{0}/sip-{0}.tar.gz'.format(args.sipver),
+                path, reporthook)
 
         with TarFile.open(path, 'r:*') as tf:
             tf.extractall(path='src')
@@ -113,9 +114,10 @@ def buildPyQt5():
                 per = 100
             print('download PyQt5_gpl-%s.tar.gz %.2f%%' % (args.pyqtver, per))
 
-        urllib.request.urlretrieve(
-            'https://www.riverbankcomputing.com/static/Downloads/PyQt5/{0}/PyQt5_gpl-{0}.tar.gz'.format(
-                args.pyqtver), path, reporthook)
+        if not os.path.isfile(path):
+            urllib.request.urlretrieve(
+                'https://www.riverbankcomputing.com/static/Downloads/PyQt5/{0}/PyQt5_gpl-{0}.tar.gz'.format(
+                    args.pyqtver), path, reporthook)
 
         with TarFile.open('src/PyQt5_gpl-{0}.tar.gz'.format(args.pyqtver), 'r:*') as tf:
             tf.extractall(path='src')
